LayoutTests/imported/w3c/web-platform-tests/css/css-position/position-absolute-table-001-expected.txt

Table css width/height are interpreted differently: they are the minimum width. Absolute positioning code should respect this.

t1
div makes cell larger.
t1
div makes cell larger.
t1
div makes cell larger.

PASS table size is interpreted as border-box width by default
PASS table size is interpreted as minimum width
PASS table size border-box
PASS table size border-box interpreted as minimum width
FAIL table size content-box assert_equals: expected 100 but got 60
PASS table size content-box interpreted as minimum width
